refactor(compartment-mapper): Pare out a test scaffold
